pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biocrate"
version = "0.1.0"
authors = [
  { name="Zhijun Zhang", email="jundecno@gmail.com" }
]
description = "A toolkit for biological and molecular data processing"
readme = "README.md"
keywords = [
  "bioinformatics",
  "molecular",
  "chemistry",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Chemistry",
]
requires-python = ">=3.10"
dependencies = [
  "numpy==1.26.4",
  "pandas>=1.3.0",
  "mlcrate>=0.1.0",
  "lmdb>=1.0.0",
  "orjson>=3.0.0",
  "rootutils>=0.1.0",
  "biopython>=1.78.0",
  "pyfastx>=0.8.0",
  "rdkit>=2023.09.1",
  "pysam>=0.20.0",
  "pybedtools>=0.9.0",
  "periodictable>=1.7.0",
]